# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meeting Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/timeline:
    get:
      operationId: getProjectTimeline
      tags: [projects]
      summary: Chronological meeting highlights for a project, with AI-generated change detection between meetings
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Project timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTimeline"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Timeline generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TimelineEntry:
      type: object
      required: [meetingId, title, date, highlight]
      properties:
        meetingId:
          type: number
        title:
          type: string
        date:
          type: string
          format: date
        highlight:
          type: string
        changeFromPrevious:
          type: ["string", "null"]
    ProjectTimeline:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEntry"
...