# lib/api-zod/src/generated/types/projectTimeline.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meeting Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineEntry } from './timelineEntry';

export interface ProjectTimeline {
  entries: TimelineEntry[];
}