# lib/api-zod/src/generated/types/timelineEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meeting Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEntry {
  meetingId: number;
  title: string;
  date: Date;
  highlight: string;
  /** @nullable */
  changeFromPrevious?: string | null;
}